Change Subscription of SpreeadsheetView_CellLeft to CellEndEdit
This is so we can signal when a user finishes editing a cell instead of
when they leave. This was causing a bug where the bottom most cell would
not get updated properly if the user pressed enter. The datagridView
would refocus the same cell and not invoke the CellLeave subscription